+++ db/migrate/20190505064000_create_memberships.rb
@@ -4,7 +4,7 @@ class CreateMemberships < ActiveRecord::Migration[5.2]
   def change
     create_table :memberships do |t|
       t.boolean :active, default: true
-      t.integer :dues
+      t.float :dues
       t.integer :user_id
 
       t.timestamps
